+++ CHANGELOG.md
@@ -20,6 +20,7 @@ As a minor extension, we have adopted a slightly different versioning convention
 - **UNSTABLE** Cardano transactions certification:
   - Optimize the performances of the computation of the proof with a Merkle map.
   - Handle rollback events from the Cardano chain by removing stale data.
+  - Preload Cardano transactions and Block Range Roots at signer & aggregator startup.
 
 - Crates versions: